requirements.txt
Django==3.2.8
django-debug-toolbar==3.2.2
django-extensions==3.1.3
Pillow==8.4.0

# development dependencies
ipython==7.28.0
black==21.9b0
